requirements.txt
httpx
anthropic
google-generativeai
numpy
rich
pyfiglet
pyreadline3
requests
pyyaml
prometheus_client
pytest
pytest-mock
flake8
mcp
